==> pipeline_stages/05_downstream_analysis/03-1_annotate_clusters.Rmd <==
---
jupyter:
  jupytext:
    text_representation:
      extension: .Rmd
      format_name: rmarkdown
      format_version: '1.0'
      jupytext_version: 0.8.6
  kernelspec:
    display_name: Python 3
    language: python
    name: python3
---

```{python}
import pandas as pd
import scanpy.api as sc
import numpy as np
from matplotlib import pyplot as plt
import sys
import os
sys.path.append("lib")
sys.path.append(os.path.abspath("."))
from jupytertools import setwd, display
setwd()
from goatools.base import download_go_basic_obo, download_ncbi_associations
from goatools.obo_parser import GODag
from goatools.associations import read_associations, read_ncbi_gene2go
from goatools.go_enrichment import GOEnrichmentStudy
import seaborn as sns
from plotnine import (
    ggplot, aes, geom_text, geom_point, facet_wrap, facet_grid, 
    ylim, theme_538, theme_bw, scale_y_reverse, xlim, geom_bar, 
    scale_color_manual
)
import itertools
from plotting import colors, plot_umap, names
```

```{python}
adata = sc.read_h5ad("results/downstream_analysis/adata_cd8_clusters.h5ad")
```

## prepare for DE analysis in R

```{python}
# prepare R objects
tmp_adata = sc.AnnData(X=adata.raw.X, obs=adata.obs, var=adata.raw.var)
sc.pp.filter_genes(tmp_adata, min_cells=300)
# edgeR expects raw counts, normalization does not have any effect. We can therefore simply undo log1p
r_mat = np.expm1(tmp_adata.X).T
r_obs = tmp_adata.obs
r_genes = pd.DataFrame().assign(gene_symbol=tmp_adata.var.index.values)
```

```{python}
import scipy.io
```

```{python}
# scipy.io.mmwrite("results/downstream_analysis/edger/r_mat.mtx", r_mat)
```

```{python}
# # np.savetxt("results/downstream_analysis/edger/r_mat.csv", r_mat)
# r_obs.to_csv("results/downstream_analysis/edger/r_obs.csv")
# r_genes.to_csv("results/downstream_analysis/edger/r_genes.csv")
```

## re-import results from R

```{python}
dataset_names = names["dataset"]
adata.obs["dataset"] = [dataset_names[d] for d in adata.obs["dataset"]]
```

```{python}
cluster_names = names["cluster"]
cluster_map = {str(i): c for i, c in enumerate(sorted(cluster_names.values()))}
adata.obs["clusters"] = [cluster_map[c] for c in adata.obs["clusters"]]
adata.uns["clusters_colors"] = np.array(list(colors["clusters"].values()))
```

```{python}
sc.tl.paga(adata, groups="clusters")
fig, ax = plt.subplots(figsize=(6, 6))
sc.pl.paga(adata, threshold=None, color="clusters", frameon=False, ax=ax)
fig.savefig("results/downstream_analysis/cd8_paga.pdf")
```

```{python}
fig, ax = plt.subplots(figsize=(9, 8))
sc.pl.draw_graph(adata, color="clusters", ax=ax, frameon=False, legend_loc="on data", 
                 legend_fontsize=12, size=8)
fig.savefig("results/downstream_analysis/cd8_paga_fa.jpg", quality=100, dpi=300)
```

```{python}
de_genes = pd.read_csv("results/downstream_analysis/edger/de_genes.tsv", sep="\t", index_col=0).\
    assign(cluster = lambda X: [cluster_names[x] for x in X["cluster"]])
clusters = np.unique(de_genes["cluster"].values)
```

```{python}
de_genes_pos = de_genes.loc[de_genes["logFC"] > 0, :]
de_genes_pos = de_genes_pos.groupby("cluster")\
    .apply(lambda df: df.assign(rank=df["logFC"].rank(ascending=False)))\
    .reset_index(drop=True)
```

```{python}
de_genes_count = de_genes_pos.query("FDR < 0.001").groupby("cluster").count().reset_index().\
    assign(label = lambda x: ["n=" + str(l) for l in x["rank"]])
```

### most-DE genes

```{python}
tmp_plt = (
    ggplot(de_genes_pos.loc[de_genes_pos["rank"] <= 10, :],
       aes(x="logFC", y="rank")) +
     geom_text(aes(label="'   ' + gene_symbol"), angle=0, ha="left", size=6) +
     geom_point(aes(color="cluster")) + 
     geom_text(aes(label="label", x=5, y=10), data=de_genes_count, ha="right", size=8) + 
     facet_wrap(' ~ cluster') + 
    theme_bw() + 
    scale_y_reverse(limits=(10, .5)) + 
    xlim(0, 5) + 
    scale_color_manual(list(colors["clusters"].values()), guide=False) 
)
tmp_plt.save("results/downstream_analysis/edger/de_top_genes.pdf", width=18, height=14, units='cm', dpi=300)
tmp_plt
```

```{python}
sc.pl.draw_graph(adata, color=["IFNG", "TNF", "IL2", "GZMA", "GZMH", "PDCD1"])
```

```{python}
genes = ["HAVCR2", "LAG3", "CXCL13", 
         "LAYN", "PRF1", "GZMB", "GZMK", 
         "ENTPD1", "ZNF683", "IFNG"]
gene_it = iter(genes)
fig, axs = plt.subplots(3, 4, figsize=(12, 9))
for row in axs:
    for ax in row:
        try:
            gene = next(gene_it)
            plot_umap(adata, rep="X_draw_graph_fa", color=gene, ax=ax, title=gene, size=.5)
        except StopIteration:
            pass
        ax.axis("off")

plt.colorbar(axs[0, 0].collections[0], ax=axs[:, 3], use_gridspec=True, fraction=0.05)
fig.savefig("results/downstream_analysis/gene_expr_fa.jpg", quality=100, dpi=300)
```

```{python}
sc.pl.dotplot(adata, genes, groupby='clusters', color_map='viridis', figsize=(3, 2.2), show=False)
plt.savefig("results/downstream_analysis/gene_expr_dotplot.pdf", bbox_inches="tight")
```

```{python}

```

```{python}
def jaccard_ind(set1, set2):
    set1 = set(set1)
    set2 = set(set2)
    return len(set1 & set2)/len(set1 | set2)
```

```{python}
jaccard = np.empty((len(clusters), len(clusters)))
for i1, i2 in itertools.product(range(len(clusters)), range(len(clusters))):
    c1, c2 = (clusters[i1], clusters[i2])
    jaccard[i1, i2] = jaccard_ind(
        de_genes_pos.loc[de_genes_pos["cluster"] == c1,"gene_symbol"].tolist(),
        de_genes_pos.loc[de_genes_pos["cluster"] == c2,"gene_symbol"].tolist()
    )
```

```{python}
sns.clustermap(jaccard, figsize=(4,4))
```

## GO-term enrichment

```{python}
obo_fname = download_go_basic_obo("results/downstream_analysis/go/go-basic.obo")
```

```{python}
geneid2gos = read_associations("tables/go_term_assoc.tsv")
```

```{python}
obodag = GODag(obo_fname)
```

```{python}
goaeobj = GOEnrichmentStudy(
    r_genes["gene_symbol"].tolist(), # list of background genes 
    geneid2gos, 
    obodag, 
    propagate_counts = False, # propagate counts to parent terms
    alpha = 0.01,
    methods = ['fdr_bh']
)
```

```{python}
goea_results_all = dict()
for cluster in clusters:
    tmp_de = de_genes_pos[(de_genes_pos["cluster"] == cluster) &
                          (de_genes_pos["rank"] <= 100)]
    goea_results_all[cluster] = goaeobj.run_study(tmp_de["gene_symbol"].tolist()) # list of genes to test. 
```

```{python}
goea_table = pd.DataFrame.from_records(
    ({
        "cluster" : cluster,
        "go": rec.GO,
        "go_name": rec.goterm.name,
        "go_namespace": rec.goterm.namespace,
        "fdr": rec.p_fdr_bh,
        "p": rec.p_uncorrected
    } for cluster, records in goea_results_all.items() for rec in records),
    columns = ["cluster", "go", "go_name", "go_namespace", "fdr", "p"]
)

```

```{python}
goea_table
```

```{python}
go_tab_fil = goea_table[(goea_table["go_namespace"] == "biological_process")]
go_tab_fil = go_tab_fil.assign(log_p = -np.log10(go_tab_fil['p']))
go_tab_pivot = go_tab_fil[["cluster", "go", "go_name", "log_p"]].pivot_table(index=["go", "go_name"], columns="cluster")
go_tab_pivot.columns = [c for _, c in go_tab_pivot.columns]
# the N most enriched terms for each cluster
top_go = [go_tab_fil.loc[(go_tab_fil["cluster"] == cluster) & 
                         (go_tab_fil["fdr"] < 0.01), :]\
                              .sort_values("fdr")["go"].tolist()[:5] 
          for cluster in clusters]
top_go = set(go for go_list in top_go for go in go_list)

```

```{python}
ax = sns.clustermap(go_tab_pivot[go_tab_pivot.index.get_level_values('go').isin(top_go)],
              col_cluster=False, figsize=(4, 9), metric="correlation")
ax.ax_heatmap.xaxis.set_ticks_position('top')
ax.ax_heatmap.xaxis.set_ticklabels(ax.ax_heatmap.xaxis.get_ticklabels(), rotation=90)
ax.ax_heatmap.set_ylabel(None)
# ax.tight_layout()
ax.savefig("results/downstream_analysis/go_biological_process.pdf")
```

```{python}
go_tab_fil = goea_table[(goea_table["go_namespace"] == "molecular_function")]
go_tab_fil = go_tab_fil.assign(log_p = -np.log10(go_tab_fil['p']))
go_tab_pivot = go_tab_fil[["cluster", "go", "go_name", "log_p"]].pivot_table(index=["go", "go_name"], columns="cluster")
go_tab_pivot.columns = [c for _, c in go_tab_pivot.columns]
# the N most enriched terms for each cluster
top_go = [go_tab_fil.loc[(go_tab_fil["cluster"] == cluster) & 
                         (go_tab_fil["fdr"] < 0.01), :]\
                              .sort_values("fdr")["go"].tolist()[:5] 
          for cluster in clusters]
top_go = set(go for go_list in top_go for go in go_list)

```

```{python}
ax = sns.clustermap(go_tab_pivot[go_tab_pivot.index.get_level_values('go').isin(top_go)],
              col_cluster=False, figsize=(4, 6.5), metric="correlation")
ax.ax_heatmap.xaxis.set_ticks_position('top')
ax.ax_heatmap.set_ylabel(None)
ax.ax_heatmap.xaxis.set_ticklabels(ax.ax_heatmap.xaxis.get_ticklabels(), rotation=90)
ax.savefig("results/downstream_analysis/go_molecular_function.pdf")
```

```{python}

```

## export for analysis in R
(the labeled adata obs)

```{python}
r_obs2 = adata.obs
r_obs2.to_csv("results/downstream_analysis/edger/r_obs2.csv")
```

```{python}
adata_all = sc.read_h5ad("results/downstream_analysis/adata_annotated_integrated.h5ad")
```

```{python}
adata_all.obs["dataset"] = [dataset_names[d] for d in adata_all.obs["dataset"]]
```

```{python}
r_obs_all = adata_all.obs
r_obs_all.to_csv("results/downstream_analysis/edger/r_obs_all.csv")
```

```{python}
adata.write_h5ad("results/downstream_analysis/adata_cd8_clusters_annotated.h5ad", compression="lzf")
```

```{python}

```
