==> pipeline_stages/05_downstream_analysis/03-2_plot-distributions.Rmd <==
---
jupyter:
  jupytext:
    text_representation:
      extension: .Rmd
      format_name: rmarkdown
      format_version: '1.0'
      jupytext_version: 0.8.6
  kernelspec:
    display_name: R
    language: R
    name: ir
---

```{r}
library(conflicted)
library(readr)
library(dplyr)
library(ggplot2)
library(cowplot)
library(dbplyr)
library(tidyr)
library(tibble)
conflict_prefer("filter", "dplyr")
conflict_prefer("select", "dplyr")
library(RSQLite)
library(ggpubr)
library(ggbeeswarm)
library(stringr)
library(cowplot)
library(ppcor)
library(ggcorrplot)
conflict_prefer("ggsave", "cowplot")
# library(ggpubr)
fig_size = function(width, height) options(repr.plot.width = width, repr.plot.height = height)
options(repr.matrix.max.cols=50, repr.matrix.max.rows=6)
```

```{r}
r_obs = read_csv("../../results/downstream_analysis/edger/r_obs2.csv", guess_max=100000)
r_obs_all = read_csv("../../results/downstream_analysis/edger/r_obs_all.csv", guess_max=1000000)
```

```{r}
colors = c(
    "BRCA"="#7fc97f",
    "LIHC"="#beaed4",
    "NSCLC"="#fdc086",
    "LUAD"="#fdc086",
    "LUSC"="#fdc086"
)
colors_cluster = c(
 'C0 - GZMK'='#e41a1c',
 'C1 - ZNF683'='#377eb8',
 'C2 - exhaustion'= '#4daf4a',
 'C3 - ZNF683/chemokine'= '#984ea3',
 'C4 - IL7R'= '#ff7f00',
 'C5 - mitotic'= '#ffff33',
 'C6 - heat shock'= '#a65628',
 'C7 - IFIT'= '#f781bf',
 'C8 - Immunoglobulin'= '#999999'
)
```

```{r}
r_obs = r_obs %>% mutate(clusters=as.factor(clusters)) %>%
  mutate(tumor_type = ifelse(tumor_type == "LUAD", "NSCLC", tumor_type)) %>%
  mutate(batch_patient = paste0(dataset, " - ", patient))
```

```{r}
r_obs_all = r_obs_all %>% 
  mutate(tumor_type = ifelse(tumor_type == "LUAD", "NSCLC", tumor_type)) %>%
  mutate(batch_patient = paste0(dataset, " - ", patient))
```

```{r}
tcga_db = dbConnect(RSQLite::SQLite(), "/db/tcga.sqlite", flags = RSQLite::SQLITE_RO)
```

```{r}
infiltration = tbl(tcga_db, "infiltration") %>% 
    inner_join(tbl(tcga_db, "sample")) %>%
    filter(cohort %in% c("BRCA", "LIHC", "LUAD", "LUSC")) %>%
    filter(method == "epic", sample_type=="TP") %>% 
    collect()
```

```{r}
cd8_infil_tcga = infiltration %>% 
    spread(cell_type, estimate) %>%
    mutate(fraction = `T cell CD8+`/(`T cell CD8+`+`T cell CD4+`)) %>%
    select(sample, fraction, cohort)
    
```

```{r}
p = cd8_infil_tcga %>% 
    ggplot(aes(x = cohort, y=fraction)) + 
        geom_quasirandom(color="grey") + 
        # scale_color_manual(values=colors, guide=FALSE) + 
        xlab("tumor type") + 
        ylab("proportion CD8+") + 
        theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
p_tcga = p %>%
        ggadd("median_mad", color="black") + coord_cartesian(ylim=c(0, 1))
```

```{r}
count_cd8 = r_obs %>% group_by(tumor_type, batch_patient) %>% summarise(n_cd8=n())
count_all = r_obs_all %>% filter(
    cell_type %in% c("T cell CD8+", "T cell CD4+", "T cell reg."),
    origin %in% c("tumor_primary", "tumor_edge")
) %>% group_by(tumor_type, dataset, batch_patient) %>% summarise(n=n())
cd8_infil_sc = count_cd8 %>% inner_join(count_all) %>% 
    mutate(fraction=n_cd8/n)
```

```{r}
# check -> does not cluster by dataset. 
fig_size(5, 2)
cd8_infil_sc %>% 
    ggplot(aes(x = tumor_type, y=fraction)) + 
                    geom_quasirandom(aes(color=dataset)) 
```

```{r}
p = cd8_infil_sc %>% 
    ggplot(aes(x = tumor_type, y=fraction)) + 
                    geom_quasirandom(aes(color=dataset)) + 
                    scale_color_brewer(type="qual", palette="Set1") + 
                    # scale_color_manual(values=colors, guide=FALSE) + 
                     coord_cartesian(ylim=c(0, 1)) + 
                     stat_compare_means(method="wilcox",
                                        label.y=c(.95, .85, .75),
                                           comparisons=list(c("BRCA", "LIHC"),
                                                            c("BRCA", "NSCLC"),
                                                            c("LIHC", "NSCLC"))) + 
                    xlab("tumor type") + 
                    ylab("proportion CD8+") + 
                    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

p_sc = p %>% ggadd("median_mad", color="black")
```

```{r}
fig_size(8, 3)
plot_grid(p_sc, p_tcga, nrow=1, align="v", rel_widths=c(10,5), labels=c("A", "B"))
ggsave("../../results/downstream_analysis/cd8_ratio.pdf", width = 8, height = 3)
```

Lung has few CD8+ T cells. This is consistent with estimates based on TCGA data. 
(-> compute the fraction of CD8+ T cells compared to other cells in tumor_primary samples) 

**This whole figure does not make sense! Some studies FACS sorted for CD4/CD8 T cells completely distorting  the ratio! **


## absolute cluster size

```{r}
fig_size(6, 2.5)
ggplot(r_obs, aes(x=clusters, fill=tumor_type)) + 
  geom_bar() +
scale_fill_manual(values=colors) + 
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) + 
        coord_flip() + 
    xlab("cluster")
```

```{r}
fig_size(3, 2.5)
r_obs %>% 
  mutate(cluster_short = str_sub(clusters, 1, 2)) %>% 
  mutate(cluster_short = factor(cluster_short, levels=sort(unique(cluster_short), decreasing = TRUE))) %>% 
    ggplot(aes(x=cluster_short, fill=clusters)) + 
  geom_bar() +
scale_fill_manual(values=colors_cluster, guide=FALSE) + 
    theme(axis.text.x = element_blank(),,
          line=element_blank()) +
    geom_text(y=100, stat='count', aes(label=..count..), rotation=90, hjust=0) + 
    xlab(NULL) + 
    coord_flip() + 
    ylab("cluster size")

ggsave("../../results/downstream_analysis/cluster_size.pdf", width = 3, height = 2.5)
```

```{r}
fig_size(3, 2.5)
r_obs %>% 
  group_by(clusters) %>% 
  summarise(n=n()) %>%
  ungroup() %>%
  mutate(frac=n/sum(n))
 
```

```{r}
cells_per_tumor_type = r_obs %>% 
  group_by(tumor_type) %>%
  summarise(cells_per_tumor_type=n())
cells_per_tumor_type
```

```{r}
all_patients = r_obs %>% select(tumor_type, batch_patient) %>% distinct()
all_clusters = r_obs %>% select(clusters) %>% distinct()
all_entities = all_patients %>% crossing(all_clusters)
```

```{r}
cluster_count_per_patient = r_obs %>% right_join(all_entities) %>% 
    group_by(tumor_type, batch_patient, clusters) %>% 
    summarise(n=sum(!is.na(index))) %>% 
    group_by(tumor_type, batch_patient) %>% 
    mutate(n_patient = sum(n)) %>% 
    ungroup() %>%
    mutate(fraction = n/n_patient)
```

```{r}
fig_size(18, 16)
cluster_count_per_patient %>% select(batch_patient, tumor_type, clusters1 = clusters, f1=fraction) %>% 
    inner_join(cluster_count_per_patient %>% select(batch_patient, tumor_type, 
                                                    clusters2 = clusters, f2=fraction)) %>%
    ggplot(aes(x=f1, y=f2)) + 
        geom_point(aes(color=tumor_type)) +
        facet_grid(clusters1~clusters2)  +
        scale_color_manual(values=colors) + 
        stat_smooth(method="lm") +
        coord_cartesian(ylim=c(0, max(cluster_count_per_patient$fraction))) + 
        stat_cor(method="pearson")
```

### partial correlations

```{r}
clusters_for_correlation = cluster_count_per_patient %>% select(batch_patient, clusters, fraction) %>% 
    spread(clusters, fraction) %>% as.data.frame() %>%
    column_to_rownames("batch_patient")

cluster_pcor = clusters_for_correlation %>% pcor(method="spearman")
cluster_cor = clusters_for_correlation %>% cor(method="spearman")
```

```{r}
fig_size(5, 4 )
ggcorrplot(cluster_cor)
```

```{r}
ggcorrplot(cluster_pcor$estimate)
```

```{r}
ggcorrplot(cluster_pcor$p.value < 0.05)
```

```{r}
cohorts = c("BRCA", "LIHC", "LUAD", "LUSC")
tcga_expr = tbl(tcga_db, "expression") %>% 
    inner_join(tbl(tcga_db, "sample")) %>% 
    filter(cohort %in% cohorts) %>%
    filter(gene_symbol %in% c("HAVCR2", "ZNF683", "CD8A", "CD8B")) %>%
    filter(sample_type == "TP") %>% 
    select(cohort, sample, gene_symbol, log2_tpm) %>%
    collect()
```

```{r}
expr_mat = tcga_expr %>% spread(gene_symbol, log2_tpm) %>% select(-cohort) %>% 
    as.data.frame() %>% column_to_rownames("sample") 
```

```{r}
expr_mat %>% ggplot(aes(x=HAVCR2/CD8A, y=ZNF683/CD8A)) + geom_point() + stat_cor() + stat_smooth(method='lm')
```

```{r}
ggcorrplot(pcor(expr_mat)$estimate)
```

```{r}
fig_size(11, 4)
p = cluster_count_per_patient %>% 
    mutate(cluster_name = str_replace_all(clusters, " - ", "\n")) %>% 
    ggplot(aes(x=tumor_type, y=fraction, color=tumor_type)) + 
    facet_wrap(~cluster_name, nrow=1) + geom_quasirandom() + 
    scale_color_manual(values=colors, guide=FALSE) + 
    stat_compare_means(method="wilcox",
                                        label.y=c(.95, .85, .75),
                                           comparisons=list(c("BRCA", "LIHC"),
                                                            c("BRCA", "NSCLC"),
                                                            c("LIHC", "NSCLC"))) + 
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
          strip.text = element_text(size=9)) + 
    coord_cartesian(ylim=c(0,1)) + 
    xlab(NULL)

p %>% ggadd("median_mad", color="black")

ggsave("../../results/downstream_analysis/cluster_distribution.pdf", width = 11, height = 4)
```

-> the variability between patients is higher than the variability between cohorts

-> need clinical information for single cell data! 

```{r}
fig_size(11, 4)
cluster_count_per_patient = r_obs %>% right_join(all_entities) %>% 
    group_by(dataset, batch_patient, clusters) %>% 
    summarise(n=sum(!is.na(index))) %>% 
    group_by(dataset, batch_patient) %>% 
    mutate(n_patient = sum(n)) %>% 
    ungroup() %>%
    mutate(fraction = n/n_patient)

p = cluster_count_per_patient %>% 
    mutate(cluster_name = str_replace_all(clusters, " - ", "\n")) %>% 
    ggplot(aes(x=dataset, y=fraction, color=dataset)) + 
    facet_wrap(~cluster_name, nrow=1) + geom_quasirandom() + 
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
          strip.text = element_text(size=9)) + 
    coord_cartesian(ylim=c(0,1)) + 
    xlab(NULL)

p %>% ggadd("median_mad", color="black")

ggsave("../../results/downstream_analysis/cluster_distribution.pdf", width = 11, height = 4)
```

```{r}

```
