==> pipeline_stages/05_downstream_analysis/03-4_signatures2.Rmd <==
---
jupyter:
  jupytext:
    text_representation:
      extension: .Rmd
      format_name: rmarkdown
      format_version: '1.0'
      jupytext_version: 0.8.6
  kernelspec:
    display_name: Python 3
    language: python
    name: python3
---

```{python}
# %load_ext autoreload
# %autoreload 2
import pandas as pd
import scanpy.api as sc
import numpy as np
from matplotlib import pyplot as plt
import sys
sys.path.append("lib")
from jupytertools import setwd, display
setwd()
from goatools.base import download_go_basic_obo, download_ncbi_associations
from goatools.obo_parser import GODag
from goatools.associations import read_associations, read_ncbi_gene2go
from goatools.go_enrichment import GOEnrichmentStudy
import seaborn as sns
from plotnine import (
    ggplot, aes, geom_text, geom_point, facet_wrap, facet_grid, 
    ylim, theme_538, theme_bw, scale_y_reverse, xlim, geom_bar,
    geom_line, theme, scale_color_brewer
)
import itertools
from plotting import colors, names
from pygenesig.gini import GiniSignatureGenerator, gini
from pygenesig.mcp_counter import MCPSignatureGenerator, MCPSignatureTester
from pygenesig.tools import translate_signatures, collapse_matrix, jaccard_ind, jaccard_mat
from pygenesig.bioqc import BioQCSignatureTester
from pygenesig.visualization import aggregate_scores, plot_score_heatmap
from collections import Counter, OrderedDict
from pprint import pprint
from toolz.functoolz import pipe as P
import matplotlib.lines as mlines
```

# import data

```{python}
adata_all = sc.read_h5ad("results/downstream_analysis/adata_annotated_integrated.h5ad")
```

```{python}
adata = sc.read_h5ad("results/downstream_analysis/adata_cd8_clusters.h5ad")
```

```{python}
cluster_names = names["cluster"]
de_genes = pd.read_csv("results/downstream_analysis/edger/de_genes.tsv", sep="\t", index_col=0).\
    assign(cluster = lambda X: [cluster_names[x] for x in X["cluster"]])
clusters = np.unique(de_genes["cluster"].values)
de_genes_pos = de_genes.loc[de_genes["logFC"] > 0, :]
de_genes_pos = de_genes_pos.groupby("cluster")\
    .apply(lambda df: df.assign(rank=df["logFC"].rank(ascending=False)))\
    .reset_index(drop=True)
```

```{python}
adata_all.obs["cell_subtype"] = adata_all.obs["cell_type"].tolist()
```

```{python}
cluster_names2 = {
    c[-1]: n for c, n in names["cluster"].items()
}
```

```{python}
adata_all.obs.loc[adata.obs_names, "cell_subtype"] = [cluster_names2[i] for i in adata.obs["clusters"].tolist()]
```

```{python}
adata_all = adata_all[adata_all.obs["cell_subtype"] != "T cell CD8+", :].copy()
```

```{python}
sc.pl.umap(adata_all, color="cell_subtype", legend_loc="on data")
```

```{python}
# adata_sub = sc.pp.subsample(adata_all, n_obs=2000, copy=True)
```

## function definitions

```{python}
def _robust_mean(series):
    series.values.sort()
    n = series.size
    min_cutoff = int(np.round(.1 * n))
    max_cutoff = int(np.round(.9 * n))
    return np.mean(series.values[min_cutoff:max_cutoff])

robust_mean = lambda df: df.apply(lambda k: _robust_mean(k), axis=1)
percentile_fun = lambda df: df.apply(lambda k: np.percentile(k, 66), axis=1)

def heatmap_precision(heatmap, reference="all"):
    """calculate precision score fore each row of aggregated heatmap. """
    precision = OrderedDict()
    if reference == "within":
        # only compare to other T cell subclusters
        heatmap = heatmap.loc[:, clusters]
    for signature in heatmap.index: 
        if reference == "outside":
            # only compare to other cell types than T cells
            tmp_heatmap = heatmap.loc[:, ~heatmap.columns.isin(set(clusters) - set([signature]))]
        else:
            tmp_heatmap = heatmap
        row_series = tmp_heatmap.loc[signature, :]
        precision[signature] = row_series[signature]/np.sum(row_series)
        
    return precision

def normalize_range(a, axis=None):
    """
    Normalize an array to values between 0 and 1.

    Parameters
    ----------
    a : np.array
        the array
    axis : int
        axis to normalize over. If None, the entire array
        will be normalized. Otherwise each row or column respectively
        will be normalized independently.

    Returns
    -------
    np.array
        the normalized array.

    """
    if axis is not None and axis > 1:
        raise ValueError("This function has only be tested until axis=1. ")

    if axis == 1:
        enum = a - np.min(a, axis=axis)[:, np.newaxis]
        denom = (np.max(a, axis=axis) - np.min(a, axis=axis))[:, np.newaxis]
    else:
        enum = a - np.min(a, axis=axis)
        denom = np.max(a, axis=axis) - np.min(a, axis=axis)
    return enum / denom
```

## prepare data for pygenesig

```{python}
target = np.array(adata_all.obs["cell_subtype"].tolist())
# expr = adata_all.raw.X[:, adata_all.raw.var_names.isin(adata_all.var_names)].T.toarray()
expr = adata_all.raw.X[:, :].T.toarray()
# expr_sub = adata_sub.raw.X.T.toarray()
# target_sub = np.array(adata_sub.obs["cell_subtype"].tolist())
# assert (adata_sub.raw.var_names == adata_all.raw.var_names).all()
```

```{python}

```

```{python}
rosetta = {i : g for i, g in enumerate(adata_all.raw.var_names)}
rosetta_inv = {g : i for i, g in enumerate(adata_all.raw.var_names)}
```

### justify cutoff

```{python}
expr_collapsed = collapse_matrix(expr, target, axis=1, aggregate_fun=robust_mean)
```

```{python}
count, cutoffs = np.histogram(expr_collapsed.values, bins=np.hstack((
    np.arange(0, 1, 0.05), np.arange(1, 2, 0.25), np.arange(2, 5, 0.5)
)), density=True)
```

```{python}
fig, ax = plt.subplots(figsize=(4, 2.5))
ax.plot(cutoffs[1:], count, marker=".")
ax.axvline(x=.25, color="black")
ax.set_yscale('log')
ax.set_ylabel("$\log_{10}$ (density)")
ax.set_xlabel("aggregated expression")
ax.set_xlim(0, 4)

fig.tight_layout()
fig.savefig("results/downstream_analysis/gini_expr_thres.pdf")
```

## standard gini


### gini on clusters only

```{python}
mask = np.isin(target, list(cluster_names.values()))
```

```{python}
sg = GiniSignatureGenerator(expr, target, min_gini=.5, max_rk=3, min_expr=0, aggregate_fun=robust_mean)
rogini_all = sg.get_rogini_format()
```

```{python}
sg2 = GiniSignatureGenerator(expr[:, mask], target[mask], min_gini=.4, max_rk=3, min_expr=0, aggregate_fun=robust_mean)
rogini_clus = sg2.get_rogini_format()
```

### grid search

```{python}
gini_range = [round(g, 2) for g in np.arange(0.05, 1.05, .05)] # in case of numpy numeric problems
```

```{python}
signatures = OrderedDict()
for gini in gini_range:
    rogini_all_fil = rogini_all.query("GINI_IDX > {} & VALUE > .25 & RANKING <= 1".format(gini))\
        .sort_values("CATEGORY")
    rogini_clus_fil = rogini_clus.query("GINI_IDX > {} & VALUE > .25 & RANKING <= 1".format(gini))\
        .sort_values("CATEGORY")
    double_rogini = rogini_all_fil.set_index(["GENEID", "CATEGORY"])\
        .join(rogini_clus_fil.set_index(["GENEID", "CATEGORY"]), how="inner", lsuffix="_2", rsuffix="_1")\
        .reset_index(drop=False)\
        .sort_values("CATEGORY")
    sig_doublegini_ind = {
        c : list() for c in cluster_names.values()
    }
    for gene_id, c in zip(double_rogini["GENEID"], double_rogini["CATEGORY"]): 
        try:
            sig_doublegini_ind[c].append(gene_id)
        except KeyError:
            pass
    if not all([len(genes) == 0 for genes in sig_doublegini_ind.values()]):
        # does not only contain empty signatures
        signatures[gini] = sig_doublegini_ind
    # print("#### gini = ", gini)
    #pprint(translate_signatures(sig_doublegini_ind, rosetta))
```

# signature testing

```{python}
mcpt = MCPSignatureTester(expr, target)
# bioqct = BioQCSignatureTester(expr_sub, target_sub)
```

```{python}
score_matrices = OrderedDict(
    (gini, mcpt.score_signatures(signature)) for gini, signature in signatures.items()
)
```

```{python}
precision_dict = OrderedDict()
for gini in signatures.keys():
    score_matrix = score_matrices[gini]
    sig_labels = mcpt.sort_signatures(signatures[gini])
    heatmap = aggregate_scores(sig_labels, score_matrix, target)
    # plot_score_heatmap(heatmap, clip=None)
    
    precision_dict[(gini, "all")] = heatmap_precision(heatmap)
    precision_dict[(gini, "within")] = heatmap_precision(heatmap, reference="within")
    precision_dict[(gini, "outside")] = heatmap_precision(heatmap, reference="outside")
    precision_dict[(gini, "n_genes")] = OrderedDict((clus, len(signatures[gini][clus])) for clus in clusters)
```

```{python}
precision_df = pd.DataFrame.from_dict(precision_dict, orient='index')
precision_df = precision_df.reset_index(drop=False)
precision_df.columns = ["gini", "prec_type"] + precision_df.columns.tolist()[2:]
```

```{python}
prec_df_melt = precision_df.melt(id_vars=["gini", "prec_type"], var_name="cluster", value_name="precision")
```

```{python}
optimal_thresholds = OrderedDict([
    ("C1 - ZNF683", 0.5),
    ("C2 - exhaustion", 0.65),
    ("C4 - IL7R", 0.4),
    ("C5 - mitotic", 0.85),
    ("C6 - heat shock", 0.6),
    ("C7 - IFIT", 0.8),
    ("C8 - Immunoglobulin", 0.85)
])
```

```{python}
fig, axs = plt.subplots(nrows=2, ncols=5, sharey=True, sharex=False, figsize=(11, 5))
it_clus = iter(clusters)
for row in range(2):
    for col in range(5):
        try:
            clus = next(it_clus)
            tmp_ax = axs[row, col]
            tmp_ax.set_xlim(0, 1)
            if col == 0:
                tmp_ax.set_ylabel("precision")
            tmp_ax.set_title(clus)
            tmp_ax.set_xlabel("gini index")
            tmp_ax.set_xticks(np.arange(0, 1.1, 0.1), minor=True)          
            tmp_ax.set_xticks(np.arange(0, 1.2, 0.2))  
            tmp_ax.set_yticks(np.arange(0, 1, 0.2))          
            tmp_ax.set_yticks(np.arange(0, 1.1, 0.1), minor=True)  
            tmp_ax.set_ylim(0, 0.9)
            tmp_ax2 = tmp_ax.twinx()
            if row == 0 and col == 4 or row ==1 and col == 3:
                tmp_ax2.set_ylabel('n_genes', color='r')
                tmp_ax2.tick_params('y', colors='r')
            else:
                tmp_ax2.yaxis.set_ticks_position('none')
                tmp_ax2.yaxis.set_ticks([])
            tmp_ax2.set_ylim(0, 22)
            if clus in optimal_thresholds:
                tmp_ax.axvline(x=optimal_thresholds[clus], linestyle="--", color="grey")
            
           
            # plot gini            
            subset = prec_df_melt.query("cluster == '{}' & prec_type=='n_genes'".format(clus))
            gini_x = subset["gini"].tolist()
            n_genes_y = subset["precision"].tolist()
            # add zero line to the end of the plot. 
            gini_x += [max(gini_x) + 0.05, 2]
            n_genes_y += [0, 0]
            l_n_genes, = tmp_ax2.plot(gini_x, n_genes_y, color="red")
            
            # plot within and outside scores
            subset = prec_df_melt.query("cluster == '{}' & prec_type=='within'".format(clus))
            l_within, = tmp_ax.plot(subset["gini"], subset["precision"], '.-', color="#1b9e77")
            
            subset = prec_df_melt.query("cluster == '{}' & prec_type=='outside'".format(clus))
            l_outside, = tmp_ax.plot(subset["gini"], subset["precision"], '.-', color="#7570b3")
            
            tmp_ax.patch.set_visible(False)
            tmp_ax.set_zorder(tmp_ax2.get_zorder() + 1)
            
        except StopIteration:
            axs[row, col].axis('off')
    

fig.tight_layout()
fig.subplots_adjust(wspace=.2, hspace=.6)
fig.legend(handles=(l_within, l_outside, l_n_genes),
    labels=("sig. unambiguity:\nCD8+ subtypes", "sig. unambiguity\nother cell types", "#genes"),
    loc=(.82, .16))

fig.savefig("results/downstream_analysis/gini_thres.pdf")
```

```{python}

```

```{python}
optimal_signatures = OrderedDict(
    (sig, signatures[gini][sig]) for sig, gini in optimal_thresholds.items()
)
```

```{python}
score_matrix = mcpt.score_signatures(optimal_signatures)
```

```{python}
sig_labels = mcpt.sort_signatures(optimal_signatures)
heatmap = aggregate_scores(sig_labels, score_matrix, target)
within_prec  = heatmap_precision(heatmap, reference="within")
outside_prec = heatmap_precision(heatmap, reference="outside")
prec_df = pd.DataFrame().assign(within=within_prec.values(), 
                                outside=outside_prec.values())
prec_df.columns = ["CD8+ subtypes", "other cell types"]

fig, (ax1, ax2, ax3) = plt.subplots(1, 3, figsize=(6.5, 4.2), gridspec_kw = {'width_ratios':[4, 4, 2]})
sns.heatmap(normalize_range(heatmap, axis=1).loc[:, clusters], ax=ax1, cbar=False, cmap='viridis')
ax1.set_title("CD8+ subtypes")
ax1.set_xlabel("")
sns.heatmap(normalize_range(heatmap, axis=1).loc[:, ~heatmap.columns.isin(clusters)], ax=ax2, cbar=False, cmap='viridis')
ax2.yaxis.set_ticks([])
ax2.set_xlabel("")
ax2.set_title("Other cell types")
sns.heatmap(prec_df, ax=ax3, cbar=False, annot=True, cmap='viridis')
ax3.yaxis.set_ticks([])
ax3.set_title("Signature\nunambiguity")

fig.tight_layout()
fig.subplots_adjust(wspace=.05)
fig.savefig("results/downstream_analysis/sig_eval.pdf")
```

```{python}
all_sig_genes = np.array([rosetta[i] for i in [g for genes in optimal_signatures.values() for g in genes]])
```

```{python}
var_group_positions = []
i = 0
for sig in optimal_signatures.values():
    var_group_positions.append((i, i+len(sig)-1))
    i += len(sig)
```

```{python}
sc.pl.dotplot(adata_all, all_sig_genes, groupby="cell_subtype", 
              var_group_positions=var_group_positions,
              var_group_labels=list(optimal_signatures.keys()), figsize=(5, 3.6), show=False, color_map='viridis')
plt.savefig("results/downstream_analysis/signature_dotplot.pdf", bbox_inches="tight")
```

```{python}
# sc.pl.heatmap(adata_all, all_sig_genes, show_gene_labels=True, group_by="cell_subtype")
```

## export signatures

```{python}
def as_tidy(sig_dict):
    for c, genes in sig_dict.items():
        for g in genes:
            yield c, g
```

```{python}
sig_df = pd.DataFrame.from_records(as_tidy(translate_signatures(optimal_signatures, rosetta)))
sig_df.columns = ["cluster", "gene_symbol"]
sig_df.to_csv("results/downstream_analysis/gini_sig.csv", index=False)
```

```{python}

```
