==> pipeline_stages/05_downstream_analysis/04_tcga.Rmd <==
---
jupyter:
  jupytext:
    text_representation:
      extension: .Rmd
      format_name: rmarkdown
      format_version: '1.0'
      jupytext_version: 0.8.6
  kernelspec:
    display_name: R
    language: R
    name: ir
---

```{r}
library(conflicted)
library(dplyr)
conflict_prefer("filter", "dplyr")
conflict_prefer("mutate", "dplyr")
conflict_prefer("select", "dplyr")
library(dbplyr)
library(tidyr)
library(tibble)
library(readr)
library(ggplot2)
library(magrittr)
library(readxl)
library(stringr)
# other packages
library(cowplot)
library(DBI)
library(survival)
library(survminer)
library(BioQC)
library(pheatmap)
library(dplyr)
library(ppcor)
library(ggcorrplot)

options(repr.matrix.max.cols=50, repr.matrix.max.rows=6)
fig_size = function(width, height) options(repr.plot.width = width, repr.plot.height = height)
```

```{r}
tcga_survival = suppressWarnings(read_xlsx("../../tables/tcga_survival.xlsx", sheet = 1))
```

```{r}
tcga_db = dbConnect(RSQLite::SQLite(), "/db/tcga.sqlite", flags = RSQLite::SQLITE_RO)
```

```{r}
indications = tbl(tcga_db, "cohort") %>% pull(cohort)
```

```{r}
cohorts = c("LIHC", "LUSC", "LUAD", "BRCA")
# cohorts = c("SKCM")
cohort_colors = c(
    "BRCA"="#7fc97f",
    "LIHC"="#beaed4",
    "LUAD"="#fdc086",
    "LUSC"="#fdc086"
)
```

```{r}
tcga = tbl(tcga_db, "expression") %>% 
    inner_join(tbl(tcga_db, "sample")) %>% 
    filter(sample_type %in% c("TP"), cohort %in% cohorts) %>% 
    inner_join(tbl(tcga_db, "infiltration")) %>% 
    filter(method == "quantiseq", cell_type == "T cell CD8+", estimate >= 0) %>% 
    select(sample, tpm=log2_tpm, gene_symbol) %>% 
    collect() %>% 
    spread(sample, tpm) %>%
    as.data.frame() %>%
    column_to_rownames("gene_symbol")
```

```{r}
gini_sig = read_csv("../../results/downstream_analysis/gini_sig.csv") 
clusters = gini_sig %>% pull(cluster) %>% unique()
signatures = sapply(clusters, function(clus) {
    gini_sig %>% filter(cluster == clus, gene_symbol %in% rownames(tcga)) %>% pull(gene_symbol)
},simplify=FALSE, USE.NAMES=TRUE)

signatures[["CD8"]] = c("CD8A", "CD8B")
signatures[["CD4"]] = c("CD4")
# signatures[["test_exhaustion"]] = c("HAVCR2")
```

```{r}
annot_df = tbl(tcga_db, "infiltration") %>% 
    filter(cell_type %in% c("T cell CD8+", "T cell CD4+"), method =="epic") %>% 
    inner_join(tbl(tcga_db, "sample")) %>%
    select(sample, cell_type, estimate, cohort) %>% 
    collect() %>% 
    spread(cell_type, estimate) %>%  
    as.data.frame() %>% 
    column_to_rownames("sample") 

annot_df = annot_df[colnames(tcga), ]
```

```{r}
mcp_score = sapply(signatures, function(sig) {
    as.matrix(tcga)[sig, ,drop=FALSE] %>% apply(2, mean)
}) %>% t()
mcp_score = mcp_score[!apply(is.nan(mcp_score), 1, any),]
```

```{r}
annot_df2 = annot_df %>% select(cohort)
annot_df2$`CD8+ score` = mcp_score["CD8", rownames(annot_df)] 
annot_df2$`CD4+ score` = mcp_score["CD4", rownames(annot_df)]
```

```{r}
ann_colors = list(cohort = cohort_colors)
```

```{r}
fig_size(8, 3)
pl = pheatmap(mcp_score[!rownames(mcp_score) %in% c("CD4", "CD8"), ], 
         clustering_distance_cols="euclidean",
         clustering_distance_rows="correlation",
         annotation_col=annot_df2[,c("cohort", "CD8+ score")], 
         show_colnames=FALSE,
         annotation_colors = ann_colors
        )

save_plot("../../results/downstream_analysis/tcga_signatures_heatmap.pdf", pl, base_width=8, base_height = 3)
```

### partial correlation

```{r}
score_mat = mcp_score[!rownames(mcp_score) %in% c("CD4"), ]
```

```{r}
# pcor_res = pcor(score_mat %>% t())
# rownames(pcor_res$estimate) = rownames(mcp_score)
# colnames(pcor_res$estimate) = rownames(mcp_score)
```

```{r}
i = 0
plotlist = lapply(cohorts, function(cohort) {
    print(i)
    tmp_mat = score_mat[, rownames(annot_df)[annot_df$cohort == cohort]]
    pl = ggcorrplot(cor(tmp_mat %>% t()), type="lower") + ggtitle(cohort)
    if (i%%2) {
        pl = pl + theme(axis.text.y=element_blank())
    }
    if (i<2) {
        pl = pl + theme(axis.text.x=element_blank())
    }
    i <<- i+1
    pl
})
```

```{r}
fig_size(8, 7)
pl = plot_grid(plotlist=plotlist, nrow=2, align = "v", axis="lb")
pl
save_plot("../../results/downstream_analysis/tcga_correlation.pdf", pl, base_width=8, base_height=7)
```

## survival

```{r}
sample = tbl(tcga_db, "sample") %>% collect()
```

```{r}
scores = mcp_score %>% 
    as_tibble(rownames="cluster") %>% 
    gather(sample, score, -cluster) %>% 
    inner_join(sample) %>% 
    inner_join(as_tibble(annot_df2, rownames="sample")) %>% 
    # should not happen, except there are two samples for the same patient 
    group_by(patient, cluster) %>% 
    summarise(score = mean(score), 
              cd4_score = mean(`CD4+ score`),
              cd8_score = mean(`CD8+ score`))
```

```{r}
tmp_surv = scores %>% 
    inner_join(tcga_survival, by=c("patient"="bcr_patient_barcode")) %>% 
    filter(type %in% cohorts) %>% 
    select(patient, cohort=type, gender, tumor_stage=ajcc_pathologic_tumor_stage, 
           birth_days_to, OS, OS.time, cluster, score, cd8_score, cd4_score) %>% 
    spread(cluster, score) %>% 
    ungroup() %>% 
    mutate(gender=as.factor(gender), tumor_stage=as.factor(tumor_stage)) 

```

## visualize using facet

```{r}
tmp_surv_status = tmp_surv %>% 
    gather(cluster, score, -patient, -cohort, -gender, -tumor_stage,
           -birth_days_to, -OS, -OS.time, -cd8_score, -cd4_score) %>% 
    filter(cluster %in% clusters) %>% 
    group_by(cluster) %>% 
    dplyr::mutate(status = if_else(score > quantile(score, 1-thres),
                            "high", if_else(score < quantile(score, thres),
                                            "low", "nd"))) %>% 
    filter(status != "nd") %>% 
    mutate(status = as.factor(status)) %>% 
    as.data.frame()


fit = surv_fit(Surv(OS.time, OS) ~ status + cohort + cluster, data=tmp_surv_status)
```

### fit coxph for each facet to obtain pvalue

```{r}
coxph_p = sapply(cohorts, function(coh) {
    sapply(clusters, function(clus) {
        tmp_tmp_surv_status = tmp_surv_status %>% 
            filter(cohort == coh, cluster == clus)
        cox_fit = coxph(Surv(OS.time, OS) ~ status + cd8_score + birth_days_to + gender, data=tmp_tmp_surv_status)
        s = summary(cox_fit)
        s_df = tbl_df(coefficients(s))
        p_values = s_df$`Pr(>|z|)`
        p_status = p_values[1]
        p_status
    })
}) %>% as_tibble(rownames="cluster") %>% gather(cohort, p_value, -cluster) %>%
mutate(p_adj = p.adjust(p_value, method="holm")) %>% 
mutate(p_adj_text = signif(p_adj, digits=2))



```

```{r}
fig_size(11, 6.5)
pl = ggsurvplot_facet(fit, data=tmp_surv_status, facet.by=c("cohort", "cluster"), 
                 pval = FALSE, color="status", palette="npg", short.panel.labs=TRUE) +
    geom_text(aes(x=3000, y=.9, label=paste0("p = ", p_adj_text)), hjust=0, data=coxph_p, size=3.5) + 
    scale_x_continuous(breaks=c(0, 5000)) + 
    xlab("OS (days)") + 
    scale_y_continuous(breaks=c(0, .4, .8))

save_plot("../../results/downstream_analysis/survival.pdf", pl, base_width = 11, base_height=6.5)
```

## ratios of clusters

```{r}
thres=.5
```

```{r}
for (clus1 in names(signatures)) {
    for(clus2 in names(signatures)) {
        if(clus1 != clus2) { 
            tmp_surv_status = tmp_surv %>% 
                dplyr::mutate(status1 = if_else(!!sym(clus1) > quantile(!!sym(clus1), 1-thres),
                                "high", if_else(!!sym(clus1) < quantile(!!sym(clus1), thres),
                                                "low", "nd"))) %>%   
                dplyr::mutate(status2 = if_else(!!sym(clus2) > quantile(!!sym(clus2), 1-thres),
                                "high", if_else(!!sym(clus2) < quantile(!!sym(clus2), thres),
                                                "low", "nd"))) %>%
                filter(status1 != "nd", status2 != "nd") %>% 
                mutate(status = as.factor(paste0("c1-", status1, '/', 'c2-', status2))) %>%
                # filter(status1 == "high" & status2 == "low" | status1 == "low" & status2 == "high") %>% 
                as.data.frame()
          
            
            fit = surv_fit(Surv(OS.time, OS) ~ status, data=tmp_surv_status)
    
            cox_fit = coxph(Surv(OS.time, OS) ~ status + cd8_score + birth_days_to + gender, data=tmp_surv_status)

            s = summary(cox_fit)
            s_df = tbl_df(coefficients(s))
            p_values = s_df$`Pr(>|z|)`
            p_status = p_values[1]

            print(ggsurvplot(fit, data=tmp_surv_status, pval = TRUE)  + 
                ggtitle(paste0(clus1, "/", clus2, ", p = ", round(p_status, 4))))
        }
    }
}

```

```{r}

```

```{r}

```
